+++ itella-cod.php
@@ -8,7 +8,7 @@
  * @wordpress-plugin
  * Plugin Name:       Smartposti COD
  * Description:       Card on delivery payment method for Smartposti shipping
- * Version:           1.0.4
+ * Version:           1.0.5
  * Author:            Itella Team
  * Author URI:        https://itella.lt/en/
  * License:           GPL-2.0+